refactor: change name of the folder
