+++ src/components/Textbook/Word/Word.ts
